fix(system): harden Process move/exec and Pipe lifecycle

+++ lib/public/StormByte/system/process.cxx
@@ -5,78 +5,110 @@
 #ifdef UNIX
 #include <sys/wait.h>
 #include <signal.h>
+#include <cstdlib>
 #else
 #include <tlhelp32.h>
+#include <sstream>
+#include <iterator>
 #endif
 
 using namespace StormByte::System;
 
-Process::Process(const std::filesystem::path& prog, const std::vector<std::string>& args):m_status(Status::RUNNING),
-m_pstdout(new Pipe()), m_pstdin(new Pipe()), m_pstderr(new Pipe()),
-m_program(prog), m_arguments(args) {
+Process::Process(const std::filesystem::path& prog, const std::vector<std::string>& args):
+	m_status(Status::RUNNING),
+#ifdef UNIX
+	m_pid(-1),
+#endif
+	m_pstdout(std::make_unique<Pipe>()),
+	m_pstdin(std::make_unique<Pipe>()),
+	m_pstderr(std::make_unique<Pipe>()),
+	m_program(prog),
+	m_arguments(args) {
+#ifdef WINDOWS
+	ZeroMemory(&m_siStartInfo, sizeof(STARTUPINFOW));
+	ZeroMemory(&m_piProcInfo, sizeof(PROCESS_INFORMATION));
+#endif
 	Run();
 }
 
-Process::Process(std::filesystem::path&& prog, std::vector<std::string>&& args):m_status(Status::RUNNING),
-m_pstdout(new Pipe()), m_pstdin(new Pipe()), m_pstderr(new Pipe()),
-m_program(std::move(prog)), m_arguments(std::move(args)) {
+Process::Process(std::filesystem::path&& prog, std::vector<std::string>&& args):
+	m_status(Status::RUNNING),
+#ifdef UNIX
+	m_pid(-1),
+#endif
+	m_pstdout(std::make_unique<Pipe>()),
+	m_pstdin(std::make_unique<Pipe>()),
+	m_pstderr(std::make_unique<Pipe>()),
+	m_program(std::move(prog)),
+	m_arguments(std::move(args)) {
+#ifdef WINDOWS
+	ZeroMemory(&m_siStartInfo, sizeof(STARTUPINFOW));
+	ZeroMemory(&m_piProcInfo, sizeof(PROCESS_INFORMATION));
+#endif
 	Run();
 }
 
-Process::Process(Process&& proc) noexcept {
-	#ifdef UNIX
-		m_pid = proc.m_pid;
-	#else
-		m_piProcInfo = proc.m_piProcInfo;
-		ZeroMemory(&proc.m_piProcInfo, sizeof(PROCESS_INFORMATION));
-		m_siStartInfo = proc.m_siStartInfo;
-		ZeroMemory(&proc.m_siStartInfo, sizeof(STARTUPINFOW));
-	#endif
-	m_pstdout = proc.m_pstdout;
-	proc.m_pstdout = nullptr;
-	m_pstdin = proc.m_pstdin;
-	proc.m_pstdin = nullptr;
-	m_pstderr = proc.m_pstderr;
-	proc.m_pstderr = nullptr;
-	m_status = proc.m_status;
-	proc.m_status = Status::TERMINATED;
-	m_program = std::move(proc.m_program);
-	m_arguments = std::move(proc.m_arguments);
+void Process::ReleaseOwnership() noexcept {
+#ifdef UNIX
+	m_pid = -1;
+#else
+	ZeroMemory(&m_piProcInfo, sizeof(PROCESS_INFORMATION));
+	ZeroMemory(&m_siStartInfo, sizeof(STARTUPINFOW));
+#endif
+	m_status = Status::TERMINATED;
+	m_pstdout.reset();
+	m_pstdin.reset();
+	m_pstderr.reset();
+	m_forwarder.reset();
+}
+
+Process::Process(Process&& proc) noexcept:
+	m_status(proc.m_status),
+#ifdef UNIX
+	m_pid(proc.m_pid),
+#else
+	m_siStartInfo(proc.m_siStartInfo),
+	m_piProcInfo(proc.m_piProcInfo),
+#endif
+	m_pstdout(std::move(proc.m_pstdout)),
+	m_pstdin(std::move(proc.m_pstdin)),
+	m_pstderr(std::move(proc.m_pstderr)),
+	m_program(std::move(proc.m_program)),
+	m_arguments(std::move(proc.m_arguments)),
+	m_forwarder(std::move(proc.m_forwarder)) {
+	proc.ReleaseOwnership();
 }
 
 Process& Process::operator=(Process&& proc) noexcept {
 	if (this != &proc) {
-		#ifdef UNIX
-			m_pid = proc.m_pid;
-		#else
-			m_piProcInfo = proc.m_piProcInfo;
-			ZeroMemory(&proc.m_piProcInfo, sizeof(PROCESS_INFORMATION));
-			m_siStartInfo = proc.m_siStartInfo;
-			ZeroMemory(&proc.m_siStartInfo, sizeof(STARTUPINFOW));
-		#endif
-		m_pstdout = proc.m_pstdout;
-		proc.m_pstdout = nullptr;
-		m_pstdin = proc.m_pstdin;
-		proc.m_pstdin = nullptr;
-		m_pstderr = proc.m_pstderr;
-		proc.m_pstderr = nullptr;
+		Wait();
 		m_status = proc.m_status;
-		proc.m_status = Status::TERMINATED;
+#ifdef UNIX
+		m_pid = proc.m_pid;
+#else
+		m_siStartInfo = proc.m_siStartInfo;
+		m_piProcInfo = proc.m_piProcInfo;
+#endif
+		m_pstdout = std::move(proc.m_pstdout);
+		m_pstdin = std::move(proc.m_pstdin);
+		m_pstderr = std::move(proc.m_pstderr);
 		m_program = std::move(proc.m_program);
 		m_arguments = std::move(proc.m_arguments);
+		m_forwarder = std::move(proc.m_forwarder);
+		proc.ReleaseOwnership();
 	}
 	return *this;
 }
 
 Process::~Process() noexcept {
 	Wait();
-	delete m_pstdout;
-	delete m_pstdin;
-	delete m_pstderr;
-	#ifdef WINDOWS
-		ZeroMemory(&m_siStartInfo, sizeof(STARTUPINFOW));
-		ZeroMemory(&m_piProcInfo, sizeof(PROCESS_INFORMATION));
-	#endif
+	m_pstdout.reset();
+	m_pstdin.reset();
+	m_pstderr.reset();
+#ifdef WINDOWS
+	ZeroMemory(&m_siStartInfo, sizeof(STARTUPINFOW));
+	ZeroMemory(&m_piProcInfo, sizeof(PROCESS_INFORMATION));
+#endif
 }
 
 Process& Process::operator>>(Process& exe) {
@@ -85,39 +117,46 @@ Process& Process::operator>>(Process& exe) {
 }
 
 std::string& Process::operator>>(std::string& data) const {
-	*m_pstdout >> data;
+	if (m_pstdout)
+		*m_pstdout >> data;
 	return data;
+}
+
+std::string& Process::Stderr(std::string& str) const {
+	if (m_pstderr)
+		*m_pstderr >> str;
+	return str;
 }
 
 std::ostream& StormByte::System::operator<<(std::ostream& os, const Process& exe) {
 	std::string data;
-	*exe.m_pstdout >> data;
+	if (exe.m_pstdout)
+		*exe.m_pstdout >> data;
 	return os << data;
 }
 
 Process& Process::operator<<(const std::string& data) {
-	*m_pstdin << data;
+	if (m_pstdin)
+		*m_pstdin << data;
 	return *this;
 }
 
 void Process::operator<<(const System::_EoF&) {
-	m_pstdin->CloseWrite();
+	if (m_pstdin)
+		m_pstdin->CloseWrite();
 }
 
 void Process::Run() {
-	#ifdef UNIX
+#ifdef UNIX
 	m_pid = fork();
 
 	if (m_pid == 0) {
-		/* STDIN: Child reads from STDIN but does not write to */
 		m_pstdin->CloseWrite();
 		m_pstdin->BindRead(STDIN_FILENO);
 
-		/* STDOUT: Child writes to STDOUT but does not read from */
 		m_pstdout->CloseRead();
 		m_pstdout->BindWrite(STDOUT_FILENO);
 
-		/* STDERR: Child writes to STDERR but does not read from */
 		m_pstderr->CloseRead();
 		m_pstderr->BindWrite(STDERR_FILENO);
 
@@ -126,25 +165,22 @@ void Process::Run() {
 		argv.push_back(const_cast<char*>(m_program.c_str()));
 		for (size_t i = 0; i < m_arguments.size(); i++)
 			argv.push_back(m_arguments[i].data());
-		argv.push_back(NULL);
-		
+		argv.push_back(nullptr);
+
 		execvp(m_program.c_str(), argv.data());
-		// If we reach here then we failed to execute the program
+		// Child must not throw across fork boundary
+		_exit(127);
+	} else if (m_pid > 0) {
+		m_pstdin->CloseRead();
+		m_pstdout->CloseWrite();
+		m_pstderr->CloseWrite();
+	} else {
+		m_status = Status::TERMINATED;
 		throw ExecutableNotFound(m_program);
 	}
-	else {
-		/* STDIN: Parent writes to STDIN but does not read from */
-		m_pstdin->CloseRead();
-
-		/* STDOUT: Parent reads from to STDOUT but does not write to */
-		m_pstdout->CloseWrite();
-
-		/* STDERR: Parent reads from to STDERR but does not write to */
-		m_pstderr->CloseWrite();
-	}
-	#else
-	ZeroMemory(&m_piProcInfo,	sizeof(PROCESS_INFORMATION));
-	ZeroMemory(&m_siStartInfo,	sizeof(STARTUPINFOW));
+#else
+	ZeroMemory(&m_piProcInfo, sizeof(PROCESS_INFORMATION));
+	ZeroMemory(&m_siStartInfo, sizeof(STARTUPINFOW));
 	m_siStartInfo.cb = sizeof(STARTUPINFOW);
 	m_siStartInfo.hStdError = m_pstderr->WriteHandle();
 	m_siStartInfo.hStdOutput = m_pstdout->WriteHandle();
@@ -160,93 +196,87 @@ void Process::Run() {
 	cmdline.push_back(L'\0');
 	LPWSTR szCmdline = cmdline.data();
 
-	if (CreateProcessW(    NULL,
-						szCmdline,            // command line 
-						NULL,                // process security attributes 
-						NULL,                // primary thread security attributes 
-						TRUE,                // handles are inherited 
-						CREATE_NO_WINDOW,    // creation flags 
-						NULL,                // use parent's environment 
-						NULL,                // use parent's current directory 
-						&m_siStartInfo,        // STARTUPINFO pointer 
-						&m_piProcInfo)) {
-		// Set the rest of handles not inheritable by other execs
+	if (CreateProcessW(NULL,
+			szCmdline,
+			NULL,
+			NULL,
+			TRUE,
+			CREATE_NO_WINDOW,
+			NULL,
+			NULL,
+			&m_siStartInfo,
+			&m_piProcInfo)) {
 		m_pstdout->WriteHandleInformation(HANDLE_FLAG_INHERIT, 0);
 		m_pstderr->WriteHandleInformation(HANDLE_FLAG_INHERIT, 0);
 		m_pstdin->ReadHandleInformation(HANDLE_FLAG_INHERIT, 0);
-	
-		// Close handles to the stdin and stdout pipes no longer needed by the child process.
-		// If they are not explicitly closed, there is no way to recognize that the child process has ended.
+
 		m_pstdout->CloseWrite();
 		m_pstderr->CloseWrite();
 		m_pstdin->CloseRead();
-	}
-	else {
+	} else {
+		m_status = Status::TERMINATED;
 		throw ExecutableNotFound(m_program);
 	}
-	#endif
+#endif
 }
 
 void Process::Send(const std::string& str) {
-	*m_pstdin << str;
+	if (m_pstdin)
+		*m_pstdin << str;
 }
 
 #ifdef UNIX
 int Process::Wait() noexcept {
-	int status = 0;
+	if (m_status == Status::TERMINATED || m_pid <= 0)
+		return -1;
 
-	// Join the forwarder thread, if any
 	if (m_forwarder) {
 		m_forwarder->join();
 		m_forwarder.reset();
 	}
 
-	// Wait for the process and handle errors
+	int status = 0;
 	if (waitpid(m_pid, &status, 0) == -1) {
 		m_status = Status::TERMINATED;
-		return -1; // Indicate failure
+		m_pid = -1;
+		return -1;
 	}
 
-	// Check if the process exited normally
-	if (WIFEXITED(status)) {
-		m_status = Status::TERMINATED;
-		return WEXITSTATUS(status); // Return the exit code
-	}
-
-	// Handle abnormal termination
 	m_status = Status::TERMINATED;
+	m_pid = -1;
+
+	if (WIFEXITED(status))
+		return WEXITSTATUS(status);
 	return -1;
 }
-
 
 pid_t Process::Pid() noexcept {
 	return m_pid;
 }
 #else
 DWORD Process::Wait() noexcept {
-	DWORD exitCode = 0;
+	if (m_status == Status::TERMINATED || m_piProcInfo.hProcess == nullptr)
+		return static_cast<DWORD>(-1);
 
-	// Join the forwarder thread, if any
 	if (m_forwarder) {
 		m_forwarder->join();
 		m_forwarder.reset();
 	}
 
-	// Wait for the process to finish
+	DWORD exitCode = 0;
 	if (WaitForSingleObject(m_piProcInfo.hProcess, INFINITE) == WAIT_FAILED) {
 		m_status = Status::TERMINATED;
-		return static_cast<DWORD>(-1); // Indicate failure
+		return static_cast<DWORD>(-1);
 	}
 
-	// Retrieve the process exit code
 	if (!GetExitCodeProcess(m_piProcInfo.hProcess, &exitCode)) {
 		m_status = Status::TERMINATED;
-		return static_cast<DWORD>(-1); // Indicate failure
+		return static_cast<DWORD>(-1);
 	}
 
-	// Clean up process handles
 	CloseHandle(m_piProcInfo.hProcess);
 	CloseHandle(m_piProcInfo.hThread);
+	ZeroMemory(&m_piProcInfo, sizeof(PROCESS_INFORMATION));
 
 	m_status = Status::TERMINATED;
 	return exitCode;
@@ -258,13 +288,15 @@ PROCESS_INFORMATION Process::Pid() {
 #endif
 
 void Process::Suspend() {
-	#ifdef UNIX
-	::kill(m_pid, SIGSTOP);
-	#else
-	HANDLE hThreadSnap = CreateToolhelp32Snapshot(TH32CS_SNAPTHREAD, 0);
-	if (hThreadSnap == INVALID_HANDLE_VALUE) {
+#ifdef UNIX
+	if (m_pid > 0)
+		::kill(m_pid, SIGSTOP);
+#else
+	if (m_piProcInfo.dwProcessId == 0)
 		return;
-	}
+	HANDLE hThreadSnap = CreateToolhelp32Snapshot(TH32CS_SNAPTHREAD, 0);
+	if (hThreadSnap == INVALID_HANDLE_VALUE)
+		return;
 
 	THREADENTRY32 te32;
 	te32.dwSize = sizeof(THREADENTRY32);
@@ -281,18 +313,20 @@ void Process::Suspend() {
 		} while (Thread32Next(hThreadSnap, &te32));
 	}
 	CloseHandle(hThreadSnap);
-	#endif
+#endif
 	m_status = Status::SUSPENDED;
 }
 
 void Process::Resume() {
-	#ifdef UNIX
-	::kill(m_pid, SIGCONT);
-	#else
-	HANDLE hThreadSnap = CreateToolhelp32Snapshot(TH32CS_SNAPTHREAD, 0);
-	if (hThreadSnap == INVALID_HANDLE_VALUE) {
+#ifdef UNIX
+	if (m_pid > 0)
+		::kill(m_pid, SIGCONT);
+#else
+	if (m_piProcInfo.dwProcessId == 0)
 		return;
-	}
+	HANDLE hThreadSnap = CreateToolhelp32Snapshot(TH32CS_SNAPTHREAD, 0);
+	if (hThreadSnap == INVALID_HANDLE_VALUE)
+		return;
 
 	THREADENTRY32 te32;
 	te32.dwSize = sizeof(THREADENTRY32);
@@ -309,59 +343,48 @@ void Process::Resume() {
 		} while (Thread32Next(hThreadSnap, &te32));
 	}
 	CloseHandle(hThreadSnap);
-	#endif
+#endif
 	m_status = Status::RUNNING;
 }
 
 void Process::ConsumeAndForward(Process& exec) {
 	m_forwarder = std::make_unique<std::thread>(
-		[&]{
-			#ifdef UNIX
+		[this, &exec] {
+#ifdef UNIX
 			std::vector<char> buffer(Pipe::MAX_READ_BYTES);
 			ssize_t bytes_read;
 			bool chunks_written = true;
 			do {
 				bytes_read = m_pstdout->Read(buffer, Pipe::MAX_READ_BYTES);
-				if (bytes_read > 0) {
-					chunks_written = exec.m_pstdin->WriteAtomic(std::string(buffer.data(), bytes_read));
-				}
+				if (bytes_read > 0)
+					chunks_written = exec.m_pstdin->WriteAtomic(std::string(buffer.data(), static_cast<size_t>(bytes_read)));
 			} while (!m_pstdout->ReadEOF() && chunks_written);
 			exec.m_pstdin->CloseWrite();
 
-			/* If chunks_written is false then it means that target executable */
-			/* already processed our input and closed connection, so we assume */
-			/* that this process is no longer needed, we send SIGTERM because  */
-			/* since we did not directly connected the IPC pipes in-out, this  */
-			/* process can't know the other one closed its stdin.              */
-			/* Furthermore, output is consumed (and ignored) to unlock current */
-			/* process in case it is locked in a pipe write operation          */
 			if (!chunks_written) {
-				kill(m_pid, SIGTERM);
-				while(!m_pstdout->ReadEOF()) {
-					std::vector<char> buffer;
-					buffer.reserve(Pipe::MAX_READ_BYTES);
-					m_pstdout->Read(buffer, Pipe::MAX_READ_BYTES);
+				if (m_pid > 0)
+					kill(m_pid, SIGTERM);
+				while (!m_pstdout->ReadEOF()) {
+					std::vector<char> discard(Pipe::MAX_READ_BYTES);
+					m_pstdout->Read(discard, Pipe::MAX_READ_BYTES);
 				}
 			}
-			#else
+#else
 			DWORD status;
 			std::vector<CHAR> buffer(Pipe::MAX_READ_BYTES);
-			SSIZE_T bytes_read;
+			DWORD bytes_read;
 			bool chunks_written = true;
 			do {
-				bytes_read = m_pstdout->Read(buffer, Pipe::MAX_READ_BYTES);
-				if (bytes_read > 0) {
+				bytes_read = m_pstdout->Read(buffer, static_cast<DWORD>(Pipe::MAX_READ_BYTES));
+				if (bytes_read > 0)
 					chunks_written = exec.m_pstdin->WriteAtomic(std::string(buffer.data(), bytes_read));
-				}
 				status = WaitForSingleObject(m_piProcInfo.hProcess, 0);
 			} while (chunks_written && status == WAIT_TIMEOUT);
-			/* See UNIX version comment above, except that in Windows we don't need */
-			/* to consume exceeding output before program can exit gracefully        */
-			if (!chunks_written) {
+
+			if (!chunks_written)
 				TerminateProcess(m_piProcInfo.hProcess, 0);
-			}
 			exec.m_pstdin->CloseWrite();
-			#endif
+#endif
 		}
 	);
 }
@@ -369,13 +392,17 @@ void Process::ConsumeAndForward(Process& exec) {
 #ifdef WINDOWS
 std::wstring Process::FullCommand() const {
 	std::stringstream ss;
-
 	std::vector<std::string> full = { m_program.string() };
 	full.insert(full.end(), m_arguments.begin(), m_arguments.end());
-	std::copy(full.begin(), full.end(), std::ostream_iterator<std::string>(ss, " "));
-	int wchars_num = MultiByteToWideChar(CP_UTF8, 0, ss.str().c_str(), -1, NULL, 0);
-	std::unique_ptr<wchar_t[]> wstr_buff = std::make_unique<wchar_t[]>(wchars_num);
-	MultiByteToWideChar(CP_UTF8, 0, ss.str().c_str(), -1, wstr_buff.get(), wchars_num);
-	return std::wstring(wstr_buff.get(), wchars_num);
+	for (size_t i = 0; i < full.size(); ++i) {
+		if (i)
+			ss << ' ';
+		ss << full[i];
+	}
+	const std::string narrow = ss.str();
+	int wchars_num = MultiByteToWideChar(CP_UTF8, 0, narrow.c_str(), -1, NULL, 0);
+	std::unique_ptr<wchar_t[]> wstr_buff = std::make_unique<wchar_t[]>(static_cast<size_t>(wchars_num));
+	MultiByteToWideChar(CP_UTF8, 0, narrow.c_str(), -1, wstr_buff.get(), wchars_num);
+	return std::wstring(wstr_buff.get());
 }
 #endif
